Move `ReadOnlyRepository` to its own package
