+++ src/main/java/ch/usi/si/seart/repository/LabelViewRepository.java
@@ -1,6 +1,7 @@
 package ch.usi.si.seart.repository;
 
 import ch.usi.si.seart.model.view.LabelView;
+import ch.usi.si.seart.repository.support.ReadOnlyRepository;
 
 public interface LabelViewRepository extends ReadOnlyRepository<LabelView, String> {
 }
