+++ src/main/java/ch/usi/si/seart/repository/LanguageStatisticsRepository.java
@@ -1,6 +1,7 @@
 package ch.usi.si.seart.repository;
 
 import ch.usi.si.seart.model.Language;
+import ch.usi.si.seart.repository.support.ReadOnlyRepository;
 
 import java.util.List;
