+++ src/main/java/ch/usi/si/seart/repository/LicenseRepository.java
@@ -1,6 +1,7 @@
 package ch.usi.si.seart.repository;
 
 import ch.usi.si.seart.model.view.License;
+import ch.usi.si.seart.repository.support.ReadOnlyRepository;
 
 public interface LicenseRepository extends ReadOnlyRepository<License, String> {
 }
