+++ src/main/java/ch/usi/si/seart/repository/TopicViewRepository.java
@@ -1,6 +1,7 @@
 package ch.usi.si.seart.repository;
 
 import ch.usi.si.seart.model.view.TopicView;
+import ch.usi.si.seart.repository.support.ReadOnlyRepository;
 
 public interface TopicViewRepository extends ReadOnlyRepository<TopicView, String> {
 }
