+++ src/main/java/ch/usi/si/seart/repository/support/ReadOnlyRepository.java
@@ -1,4 +1,4 @@
-package ch.usi.si.seart.repository;
+package ch.usi.si.seart.repository.support;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.repository.NoRepositoryBean;
 import org.springframework.data.repository.Repository;
 
-import java.io.Serializable;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,7 +19,7 @@ import java.util.Optional;
  * @author dabico
  */
 @NoRepositoryBean
-public interface ReadOnlyRepository<T, ID extends Serializable> extends Repository<T, ID> {
+public interface ReadOnlyRepository<T, ID> extends Repository<T, ID> {
 
     /*
      * (non-Javadoc)
